Refactored some logic into class Correlation
